+++ lib/core/Tiki/Command/MailQueueSendCommand.php
@@ -1,9 +1,9 @@
 <?php
-// (c) Copyright 2002-2013 by authors of the Tiki Wiki CMS Groupware Project
+// (c) Copyright 2002-2016 by authors of the Tiki Wiki CMS Groupware Project
 //
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
 // Licensed under the GNU LESSER GENERAL PUBLIC LICENSE. See license.txt for details.
-// $Id: MailInPollCommand.php 50599 2014-03-31 21:22:59Z lphuberdeau $
+// $Id: MailQueueSendCommand.php 62176 2017-04-10 06:01:52Z drsassafras $
 
 namespace Tiki\Command;
 
@@ -28,35 +28,42 @@ class MailQueueSendCommand extends Command
     {
       require_once ('lib/mail/maillib.php');
       global $prefs;
-      $logslib = TikiLib::lib('logs'); 
+      $logslib = TikiLib::lib('logs');
       tiki_mail_setup();
       $output->writeln('Mail queue processor starting...');
 
-        $messages = \TikiDb::get()->fetchAll('SELECT messageId, message FROM tiki_mail_queue');
+      $messages = \TikiDb::get()->fetchAll('SELECT messageId, message FROM tiki_mail_queue');
 
       foreach ( $messages as $message ) {
 
           $output->writeln('Sending message '.$message['messageId'].'...');
           $mail = unserialize($message['message']);
+		  $error = '';
 
-          if ($mail) {
+          if ($mail && get_class($mail) === 'Zend\Mail\Message') {
             try {
-                $mail->send();
+                tiki_send_email($mail);
                 $title = 'mail';
-            } catch (Zend_Mail_Exception $e) {
+            } catch (\Zend\Mail\Exception\ExceptionInterface $e) {
                 $title = 'mail error';
+				$error = $e->getMessage();
             }
 
-            if ($title == 'mail error' || $prefs['log_mail'] == 'y') {
-            	foreach ($mail->getRecipients() as $u) {
-            		$logslib->add_log($title, $u . '/' . $mail->getSubject());
-            	}
+            if ($error || $prefs['log_mail'] == 'y') {
+                foreach($mail->getTo() as $destination){
+                    $logslib->add_log($title, $error . "\n " . $destination->getEmail() . '/' . $mail->getSubject());
+                }
+                foreach($mail->getCc() as $destination){
+                    $logslib->add_log($title, $error . "\n " . $destination->getEmail() . '/' . $mail->getSubject());
+                }
+                foreach($mail->getBcc() as $destination){
+                    $logslib->add_log($title, $error . "\n " . $destination->getEmail() . '/' . $mail->getSubject());
+                }
             }
 
-            if ($title == 'mail error') {
+            if ($error) {
             	$query = 'UPDATE tiki_mail_queue SET attempts = attempts + 1 WHERE messageId = ?';
-            	$output->writeln('Failed.');
-            	print_r($mailer->errors);
+            	$output->writeln('Failed sending mail object id: ' . $message['messageId'] . ' (' . $error . ')');
             } else {
             	$query = 'DELETE FROM tiki_mail_queue WHERE messageId = ?';
             	$output->writeln('Sent.');
@@ -64,7 +71,7 @@ class MailQueueSendCommand extends Command
 
             \TikiDb::get()->query($query, array($message['messageId']));
           } else {
-              $output->writeln('ERROR: Unable to unserialize the mailer object.');
+              $output->writeln('ERROR: Unable to unserialize the mail object id:' . $message['messageId']);
           }
       }
       $output->writeln('Mail queue processed...');
